fix: more tests cleanup
