+++ .github/workflows/push-actions.yml
@@ -31,7 +31,7 @@ jobs:
     strategy:
       matrix:
         node_cimg_tag: [ 14, 16, 18, 20 ]
-        npm_task: ['test:unit', 'test:component', 'test:integration:version']
+        npm_task: ['test:unit', 'test:component', 'test:instrumentations']
     steps:
       - uses: actions/checkout@v3
       - name: Set up Node
