--- jest.component.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  testMatch: ['**/component/**/*.test.js'],
-  roots: ['./test'],
-  setupFilesAfterEnv: ['./jest.component.setup.js', 'jest-json'],
-};
