+++ package.json
@@ -9,15 +9,11 @@
     "build": "rm -rf dist distro.tgz && tsc --build --force && tracer=\"$(npm pack)\" && mv \"$tracer\" distro.tgz",
     "prepublishOnly": "npm run prettier:fix && npm run lint && npm run test:unit",
     "lint": "eslint \"src/**/*.{ts,tsx,js}\"",
-    "install-component": "bash -c 'source scripts/install_component.sh && install_component_dependencies'",
-    "install-integration": "bash -c 'source scripts/install_component.sh && install_integration_dependencies'",
     "prettier:ci": "prettier --list-different \"src/**/*.[jt]s\" --parser typescript",
     "prettier:fix": "prettier --write \"./src/**/*.[jt]s\" --parser typescript",
     "semantic-release": "semantic-release",
     "test:unit": "npm run build && NODE_OPTIONS=--unhandled-rejections=strict jest -c jest.unit.config.js --detectOpenHandles --runInBand --forceExit --silent",
-    "test:component": "npm run build && npm run install-component && jest -c jest.component.config.js --detectOpenHandles --runInBand --forceExit",
-    "test:integration:version": "npm run build && npm run install-integration && jest -c jest.integration.config.js --detectOpenHandles --runInBand --forceExit --silent=false",
-    "local:test:integration:version": "npm run build && npm run install-integration && jest -c jest.integration.config.js --detectOpenHandles --forceExit --silent=false",
+    "test:instrumentations": "npm run build && jest -c jest.instrumentations.config.js --detectOpenHandles --runInBand --forceExit --silent=false",
     "tsc": "tsc"
   },
   "repository": "git@github.com:lumigo-io/opentelemetry-js-distro.git",
