+++ test/instrumentations/express/app/app.js
@@ -1,4 +1,3 @@
-const axios = require('axios');
 const express = require('express');
 const bodyParser = require('body-parser');
 const app = express();
