+++ test/instrumentations/http/spans/spans-http-otel-attr.json
@@ -1,0 +1,2 @@
+{"traceId":"6d44c2fd4f3ffcbba54bf70e6dbb250f","parentId":"6c3241e57fa9a3af","name":"GET","id":"17b949beb8c1f185","kind":2,"timestamp":1683192212513000,"duration":10174,"attributes":{"http.url":"htt","http.method":"GET","http.target":"/se","net.peer.name":"loc","http.host":"loc","http.request.headers":"{\"a","http.request.body":"\"\"","net.peer.ip":"::1","net.peer.port":9000,"http.response_content_length_uncompressed":425273,"http.status_code":200,"http.status_text":"OK","http.flavor":"1.1","net.transport":"ip_","http.response.headers":"{\"c","http.response.body":"\"[{"},"status":{"code":0},"events":[],"resource":{"attributes":{"service.name":"http","telemetry.sdk.language":"nodejs","telemetry.sdk.name":"opentelemetry","telemetry.sdk.version":"1.9.0","process.pid":52907,"process.executable.name":"node","process.command":"/Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.command_line":"/opt/homebrew/Cellar/node/19.8.1/bin/node /Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.runtime.version":"19.8.1","process.runtime.name":"nodejs","process.runtime.description":"Node.js","lumigo.distro.version":"1.19.1","framework":"node","process.environ":"{\"MallocNanoZone\":\"0\",\"COLOR\":\"0\",\"npm_config_noproxy\":\"\",\"OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT\":\"3\"}"}}}
+{"traceId":"6d44c2fd4f3ffcbba54bf70e6dbb250f","name":"GET","id":"6c3241e57fa9a3af","kind":1,"timestamp":1683192212494000,"duration":37637,"attributes":{"http.url":"htt","http.host":"loc","net.host.name":"loc","http.method":"GET","http.scheme":"htt","http.target":"/la","http.user_agent":"axi","http.flavor":"1.1","net.transport":"ip_","lumigo.execution_tags.foo":"bar","lumigo.execution_tags.date":1234567,"net.host.ip":"::1","net.host.port":60368,"net.peer.ip":"::1","net.peer.port":60369,"http.status_code":200,"http.status_text":"OK"},"status":{"code":0},"events":[],"resource":{"attributes":{"service.name":"http","telemetry.sdk.language":"nodejs","telemetry.sdk.name":"opentelemetry","telemetry.sdk.version":"1.9.0","process.pid":52907,"process.executable.name":"node","process.command":"/Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.command_line":"/opt/homebrew/Cellar/node/19.8.1/bin/node /Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.runtime.version":"19.8.1","process.runtime.name":"nodejs","process.runtime.description":"Node.js","lumigo.distro.version":"1.19.1","framework":"node","process.environ":"{\"MallocNanoZone\":\"0\",\"COLOR\":\"0\",\"npm_config_noproxy\":\"\",\"OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT\":\"3\"}"}}}
