+++ test/instrumentations/http/spans/spans-http-span-attr.json
@@ -1,0 +1,2 @@
+{"traceId":"911f8b9762fa07f83b65bd91ecd3cd6a","parentId":"551a8c570fdb4d56","name":"GET","id":"13dd3788a0dc0cbf","kind":2,"timestamp":1683192210994000,"duration":13750,"attributes":{"http.url":"h","http.method":"G","http.target":"/","net.peer.name":"l","http.host":"l","http.request.headers":"{","http.request.body":"\"","net.peer.ip":":","net.peer.port":9000,"http.response_content_length_uncompressed":97,"http.status_code":200,"http.status_text":"O","http.flavor":"1","net.transport":"i","http.response.headers":"{","http.response.body":"\""},"status":{"code":0},"events":[],"resource":{"attributes":{"service.name":"http","telemetry.sdk.language":"nodejs","telemetry.sdk.name":"opentelemetry","telemetry.sdk.version":"1.9.0","process.pid":52859,"process.executable.name":"node","process.command":"/Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.command_line":"/opt/homebrew/Cellar/node/19.8.1/bin/node /Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.runtime.version":"19.8.1","process.runtime.name":"nodejs","process.runtime.description":"Node.js","lumigo.distro.version":"1.19.1","framework":"node","process.environ":"{\"OTEL_SPAN_ATTRIBUTE_VALUE_LENGTH_LIMIT\":\"1\",\"npm_config_noproxy\":\"\"}"}}}
+{"traceId":"911f8b9762fa07f83b65bd91ecd3cd6a","name":"GET","id":"551a8c570fdb4d56","kind":1,"timestamp":1683192210980000,"duration":32003,"attributes":{"http.url":"h","http.host":"l","net.host.name":"l","http.method":"G","http.scheme":"h","http.target":"/","http.user_agent":"a","http.flavor":"1","net.transport":"i","lumigo.execution_tags.foo":"f","lumigo.execution_tags.baz":true,"net.host.ip":":","net.host.port":60364,"net.peer.ip":":","net.peer.port":60366,"http.status_code":200,"http.status_text":"O"},"status":{"code":0},"events":[],"resource":{"attributes":{"service.name":"http","telemetry.sdk.language":"nodejs","telemetry.sdk.name":"opentelemetry","telemetry.sdk.version":"1.9.0","process.pid":52859,"process.executable.name":"node","process.command":"/Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.command_line":"/opt/homebrew/Cellar/node/19.8.1/bin/node /Users/mmanciop/git/opentelemetry-js-distro/test/instrumentations/http/app/app.js","process.runtime.version":"19.8.1","process.runtime.name":"nodejs","process.runtime.description":"Node.js","lumigo.distro.version":"1.19.1","framework":"node","process.environ":"{\"OTEL_SPAN_ATTRIBUTE_VALUE_LENGTH_LIMIT\":\"1\",\"npm_config_noproxy\":\"\"}"}}}
