+++ test/instrumentations/http/test-resources/large-response.json
