--- test/testUtils/spanUtils.ts
@@ -1,4 +1,0 @@
-
-export function getSpanByKind(spans: [], spanKindValue: number) {
-    return spans.find((span) => span.kind === spanKindValue);
-}
